create separate git project file
