+++ src-tauri/src/git/mod.rs
@@ -1,3 +1,4 @@
 pub mod project_folder;
 pub mod git_files;
 pub mod git_folders;
+pub mod git_project;
